requirements.txt
yfinance
pandas
yfinance
pandas
requests
beautifulsoup4
python-dotenv
schedule
openpyxl

google-generativeai
markdown

matplotlib
python-bcb
jinja2
